resizing-hbox: Don't hold a reference to removed actors
MxOffscreen holds a reference to the actor it's cloning, so save the
texture after the initial update of the offscreen and replace it with
a standard texture, so the actor can be destroyed before the animation
completes.

This fixes re-adding the same menu options to ones you just removed before
the animation finishes.