models: introduce view-model class
The ViewModel is responsible to display a portion of an underlaying
model (let's say the first 8 elements for example), it's a
view-on-a-model. The purpose to introduce such a class is to reduce
the complexity of the current Proxy class which handles the UI
elements creation as well as this view-on-a-model feature.