mex-tileout-effect: Use clutter_effect_queue_rerun
If we have a recent version of Clutter then we can avoid repainting
the whole actor when only the progress property changes by calling
clutter_effect_queue_rerun. This will cause ClutterOffscreenEffect to
bypass painting the actor and jump straight to the paint_target method
unless something else queues a redraw of the actor in the meantime.