content: introduce open method on content
The purpose of this new method is to seamlessly manipulate
directories/albums/containers of any kind of content and actual
contents.

For a content the default behavior is to open the player and access
the content. For a container the behavior will be to go further the
hierachy of the current browsing mode.