log: Pimp up our log system
We had a very simple log system to do "printf debugging". Let's import
the one from grilo (I've heavily contributed to it, not to say have
written it).

This will allow us to use the same logging mechanism both in the core
library and in the plugins we dlopen() at run time. Plus, it has several
degree of verbosity we can use too.

Fixes: https://github.com/media-explorer/media-explorer/issues/31