mmkeys: key_grab_active is TRUE/FALSE when the dbus call has been issued
This flag is only used internally to not make the dbus call if not
needed. Change it so its value is updated as soon as the dbus call is
made to grab/ungrab the keys (and reset the flag in case of error from
dbus). This allow not to issue severial calls while waiting for the dbus
answer.