shell: Remove the cursor hiding from the library
This is really something that belong either to the shell or to some
either level object in the library.

It's a better idea to do it in the notify::fullscreen-set signal handler
on the stage instead of realying that everyones calls
mex_set_fullscreen(). Actually the window manager could request
fullscreen for the application without us having to say anything.
Listening to fullscreen-set changes should cover that case.