webremote: Add a fallback for getting the mex_data_dir and crash less
I can be tricky if you're in a werid enviroment to get the correct data
directory, so for now a DEFINE is used to fall back on from the normal
way of getting the location from Mex. Additionally we try to catch the
case where the data dir has not been found so that we don't segfault.