build: Fix the out of tree build of the debug plugin
When doing an out of tree build, the generation of the mex-debug file
happens first but $(buildir)/plugins/debug does not exist first.

Create the debug directory if needed then.

Also the dependency on mex-debug.in is better expressed if it
consistently have $(srcdir) rather than relying on an implicit behaviour
of automake.