mex: introduce MexBackground & MexBackgroundManager
The MexBackgroundManager is the component responsible to manage the
background animation. Each plugin is free to register it's own actor
implementing the MexBackground interface, the MexBackgroundManager
will pick one and use it as background animation in the UI.

Currently the manager will just pick the first available component. In
the future a UI could be added to select which component should be
used.