column: split column/header
Currently each Column is a embedded in a ScrollView which is a bit of
a pain because the header of the Column is supposed to be always on
the top over the items of the Column. We already had a workaround for
this, but there is still the issue of ensuring an item is visible
within the ScrollView, which cannot work correctly because of the
header being painted on top.

So to make the whole thing simpler, we now split each Column in 2
widgets. MexColumnView contains a header + ScrollView(MexColumn) and
MexColumn contains all the items.