player: Remove the misleading priv->position variable
This variable was only used as the position to resume the content from
and never to store the position of the current stream.

Yet, a few places in the code updated that variable without any actual
meaningful effect.

So let's remove this variable altogether and compute the resume position
when we need it.