player: Put the queue resume logic next where it belongs
So, we don't resume a video from the last position if it comes from the
queue. The way it works now is by checking the "is the content being
played from the queue" condition when we query the duration and then,
later on, we check if the duration is 0

That can be way more direct by checking the queue condition when it's
time to resume, so people reading the code don't have to go through an
indirection. Also this allows to use priv->duration always instead of
the undocumented behaviour that priv->duration is 0.0 when the content
is in the queue.