generic-model: Replace our sorted insertion
Let's just write a regular binary search in an array. If found, return
the position of the found element and we can insert there.

If not found, we encode the 'first' variable into the negative number we
return. This can be used to insert the element where it belongs.

With this the '/core/model/sorted-insertion' test passes.