dvb: Import w_scan version 20110702
Ok, let's start the journey towards native DVB support by importing a
scanner, one of those that can actually scan without initial tuning data
(ie not the linux-dvb-utils one).

The goal is to make a daemon out of w_scan, but I'm importing an
unmodified copy first to be able to track all the modificatins I make to
the package.

This daemon's license is GPLv2+.