dvb: Add an Error signal to the dbus interface
When the w_scan thread calls fatal, something really bad happened. Send
a signal back to the application and exit the daemon (because we are in
an error state that cannot be recovered). The next D-Bus call will start
the daemon again.

On the client side, it means that on error, you have to start the whole
sequence again (for instance, SetCountry(); StartScan())