w_scan: Use fatal() instead of error() + exit()
When used as a daemon, I don't really want w_scan to call exit() itseft
as we need to report errors through D-Bus. For that, w_scan has a
fatal() macro and the error messages should go through there.

There was a fatal error message remaining in the code that did not use
fatal() but called exit directly, fix it.